% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_io_credentials.R
\name{adobe_target_api_headers}
\alias{adobe_target_api_headers}
\title{Format standard headers for Adobe Target API calls}
\usage{
adobe_target_api_headers(api.version = "v3")
}
\description{
Format standard headers for Adobe Target API calls
}
\examples{
adobe_target_api_headers()
}
