% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_io_credentials.R
\name{checkRenviron}
\alias{checkRenviron}
\title{Check for requisite .Renviron variables; alert if missing}
\usage{
checkRenviron(var)
}
\arguments{
\item{var}{(character) name of variable}

\item{var}{}
}
\value{
Error message if .Renviron not set
}
\description{
Check for requisite .Renviron variables; alert if missing
}
\examples{
checkRenviron("API_KEY")
}
