% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangeJWTforBearerToken.R
\name{exchangeJWTforBearerToken}
\alias{exchangeJWTforBearerToken}
\title{Exchange JWT for Bearer token (POST)
Wrapper for Adobe IO API call}
\usage{
exchangeJWTforBearerToken(path.to.config, path.to.pem)
}
\value{
Bearer token that works for 24 hours and is to be used in all the Target API calls
}
\description{
Exchange JWT for Bearer token (POST)
Wrapper for Adobe IO API call
}
\examples{
exchangeJWTforBearerToken("C://users/amanda.ludden//usermanagement.config", "C://users/amanda.ludden/secret.pem")
}
