% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivityChangelog.R
\name{extractChangelogTiming}
\alias{extractChangelogTiming}
\title{Extract start-/stop-related details from getActivityChangelog() output}
\usage{
extractChangelogTiming(changelog)
}
\arguments{
\item{changelog}{}
}
\value{
data.frame
}
\description{
Extract start-/stop-related details from getActivityChangelog() output
}
\examples{
extractChangelogTiming(getActivityChangelog("12345"))
}
