% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{getAbActivityById}
\alias{getAbActivityById}
\title{Get AB Activity by ID}
\usage{
getAbActivityById(activityId, encoding = "UTF-8")
}
\arguments{
\item{activityId}{}
}
\value{
json; see \href{https://developers.adobetarget.com/api/#get-ab-activity-by-id}{Sample Response}
}
\description{
Fetch the current definition of an AB activity if it is found as referenced by the id.
\url{http://developers.adobetarget.com/api/#get-ab-activity-by-id}
}
\examples{
getAbActivityById("298487")
}
