% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getAbPerformanceReport}
\alias{getAbPerformanceReport}
\title{Get AB Performance Report}
\usage{
getAbPerformanceReport(activityId, encoding = "UTF-8")
}
\arguments{
\item{activityId}{}
}
\value{
json; see \href{https://developers.adobetarget.com/api/#get-ab-performance-report}{Sample Response}
}
\description{
Retrieve the performance report data for the AB activity referenced by the provided id.
\url{https://developers.adobetarget.com/api/#get-ab-performance-report}
}
\examples{
getAbPerformanceReport("12345")
}
