% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{getAbReportInterval}
\alias{getAbReportInterval}
\title{Get AB Report Interval from Adobe Target getAbPerformanceReport()}
\usage{
getAbReportInterval(activityId)
}
\arguments{
\item{activityId}{}
}
\description{
Wrapper for getAbPerformanceReport adobeTaRget/Adobe Target API call (specifically reportInterval value)
Calculates duration for later
}
\examples{
getAbReportInterval("224920")
}
