% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivityChangelog.R
\name{getActivityChangelog}
\alias{getActivityChangelog}
\title{getActivityChangelog}
\usage{
getActivityChangelog(activityId, activities = "")
}
\arguments{
\item{activityId}{(character) activityId returned by listActivities()}

\item{activities}{output of listActivities}
}
\value{
data.frame of a single activity's events
}
\description{
getActivityChangelog
}
\examples{
getActivityChangelog("123456"); if only one supplied, you get a list instead of df
}
