% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiences.R
\name{getAudienceById}
\alias{getAudienceById}
\title{Get Audience by ID}
\usage{
getAudienceById(audienceId)
}
\arguments{
\item{audienceId}{}
}
\value{
json; see [Sample Response](https://developers.adobetarget.com/api/#get-audience-by-id)
}
\description{
Get the audience definition specified by the provided id.
https://developers.adobetarget.com/api/#get-audience-by-id
}
\examples{
getAudienceById("3245850")
}
