% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offers.R
\name{getOfferById}
\alias{getOfferById}
\title{Get Offer by ID}
\usage{
getOfferById(offerId)
}
\arguments{
\item{offerId}{}
}
\value{
json; see \href{https://developers.adobetarget.com/api/#get-offer-by-id}{Sample Response}
}
\description{
Retrieves the contents of an offer given an offer id (usually HTML)
\url{https://developers.adobetarget.com/api/#get-offer-by-id}
}
\examples{
getOfferById("483105")
}
