% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiences.R
\name{listAudiences}
\alias{listAudiences}
\title{List Audiences}
\usage{
listAudiences()
}
\value{
data.frame of all audiences; see [Sample Response](https://developers.adobetarget.com/api/#list-audiences)
}
\description{
List all available audiences with options to filter and sort by each available field.
http://developers.adobetarget.com/api/#list-audiences
}
\examples{
listAudiences()
}
