% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offers.R
\name{listOffers}
\alias{listOffers}
\title{List Offers}
\usage{
listOffers()
}
\value{
data.frame of all activities; see \href{https://developers.adobetarget.com/api/#list-offers}{Sample Response}
}
\description{
Retrieve the list of previously-created content offers.
\url{https://developers.adobetarget.com/api/#list-offers}
}
\examples{
listOffers()
}
