% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_io_credentials.R
\name{make_exchange_jwt_config}
\alias{make_exchange_jwt_config}
\title{Create temporary .config file for ExchangeJWT.py using variables set in adobe_io_credentials()}
\usage{
make_exchange_jwt_config()
}
\description{
Create temporary .config file for ExchangeJWT.py using variables set in adobe_io_credentials()
}
\examples{
make_exchange_jwt_config()
}
