% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSiteCatalyst.R
\name{QueueA4TReport}
\alias{QueueA4TReport}
\title{Queue A4T Report}
\usage{
QueueA4TReport(activityId,
  reportsuite.id = Sys.getenv("ADOBE_REPORTSUITEID"),
  metrics = "uniquevisitors", elements = c("activity", "experience"),
  date.granularity = "year", reportMeta = TRUE, ...)
}
\arguments{
\item{activityId}{character}

\item{reportsuite.id}{See RSiteCatalyst::QueueTrended();}

\item{metrics}{See RSiteCatalyst::QueueTrended() ;}

\item{elements}{See RSiteCatalyst::QueueTrended() ;}

\item{date.granularity}{See RSiteCatalyst::QueueTrended(); Defaults to year since that's the highest level of aggregation; if your test spans two calendar years, you'll have to aggregate}

\item{reportMeta}{boolean. If TRUE, appends duration and id to results of RSiteCatalyst::QueueTrended()}

\item{...}{additional arguments to RSiteCatalyst::QueueTrended()}
}
\description{
Uses RSiteCatalyst::QueueTrended() to get metrics for an Adobe Target AB activity that have Adobe Analytics as the reporting source (A4T)\cr
}
\examples{
# Error if activity doesn't report to Adobe Analytics 
\dontrun{
QueueAbReport(activityId = "297037", reportMeta=FALSE)
}

QueueA4TReport(activityId = "298487", reportMeta=FALSE) # reports uniquevisitors by default
QueueA4TReport(activityId = "298487", metrics = c("uniquevisitors", "visits"), reportMeta=FALSE) # or specify additional metrics, e.g., conversion
QueueA4TReport(activityId = "298487", reportMeta=TRUE) # reports uniquevisitors by default
}
