% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeAdobeNotifications.R
\name{scrapeAdobeNotifications}
\alias{scrapeAdobeNotifications}
\title{[DEPRECATED] Scrape Adobe Experience Cloud notification feed for start and complete times}
\usage{
scrapeAdobeNotifications(useDocker = TRUE,
  feed_scraper_js = dir(path.package("adobeTaRget"), pattern = "feed2.js",
  full.names = TRUE))
}
\arguments{
\item{useDocker}{boolean}

\item{feed_scraper_js}{path to javascript file to scrape feed}
}
\value{
data.frame
}
\description{
NOTE: The feed doesn't pick up on activity name changes because there is no unique ID persisted
Only try this if you can't get the System Admin privileges to use the API
}
\examples{
tests <- testCalendar(useDocker=FALSE)
}
