% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_io_credentials.R
\name{set_adobe_io_credentials}
\alias{set_adobe_io_credentials}
\title{Install Adobe IO credentials in a .config file and .Renviron for repeated use}
\usage{
set_adobe_io_credentials(ims_host, ims_endpoint_jwt, org_id, api_key,
  client_secret, tech_acct_id, priv_key_filename, ADOBE_TENANT_NAME,
  ADOBEIO_CONFIG_PATH, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{ims_host}{}

\item{ims_endpoint_jwt}{}

\item{org_id}{}

\item{api_key}{}

\item{client_secret}{}

\item{priv_key_filename}{}

\item{ADOBE_TENANT_NAME}{}

\item{ADOBEIO_CONFIG_PATH}{}

\item{tech_acct}{}
}
\description{
Install Adobe IO credentials in a .config file and .Renviron for repeated use
}
\examples{
set_adobe_io_credentials(ims_host = "xxx", 
ims_endpoint_jwt = "xxx",
org_id = "xxx",
api_key = "xxx",
client_secret = "xxx",
tech_acct = "xxx",
priv_key_filename = "xxx",
ADOBE_TENANT_NAME = "xxx",
ADOBEIO_CONFIG_PATH = "xxx"
)

set_adobe_io_credentials(ims_host = "asdf", 
ims_endpoint_jwt = "xxx",
org_id = "xxx",
api_key = "xxx",
client_secret = "xxx",
tech_acct = "xxx",
priv_key_filename = "xxx",
ADOBE_TENANT_NAME = "xxx",
ADOBEIO_CONFIG_PATH = "xxx",
overwrite=TRUE,
install=TRUE)
}
