% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offers.R
\name{updateOfferById}
\alias{updateOfferById}
\title{Update Offer by ID}
\usage{
updateOfferById(offerId, offerName, updatedContent, api.version = "v2")
}
\arguments{
\item{offerId}{}
}
\description{
Updates the content offer referenced by the id specified in the URL.
\url{https://developers.adobetarget.com/api/#update-offer-by-id}
}
\examples{
updateOfferById(offerId="535745", offerName="asdf", updatedContent="<b>asdf</b>")
}
