% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_io_credentials.R
\name{exchangeJWTforBearerToken}
\alias{exchangeJWTforBearerToken}
\title{Exchange JWT for Bearer token (POST)
Wrapper for Adobe IO API call
Writes Bearer token to the config file; write bearer token to renviron}
\usage{

  exchangeJWTforBearerToken(secret.pem.path = Sys.getenv("adobe_io_priv_key_filename"))
}
\arguments{
\item{secret.pem.path}{}
}
\value{
Bearer token that works for 24 hours needed for all Target API calls
}
\description{
Exchange JWT for Bearer token (POST)
Wrapper for Adobe IO API call
Writes Bearer token to the config file; write bearer token to renviron
}
\examples{
exchangeJWTforBearerToken()
}
