% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbPerformanceReport.R
\name{getAbPerformanceReport}
\alias{getAbPerformanceReport}
\title{Get AB Performance Report
Wrapper for Get AB Performance Report API call}
\usage{
getAbPerformanceReport(activityId)
}
\arguments{
\item{activityId}{}
}
\value{
json
}
\description{
Get AB Performance Report
Wrapper for Get AB Performance Report API call
}
\examples{
getAbPerformanceReport("12345")
}
