% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAudienceById.R
\name{getAudienceById}
\alias{getAudienceById}
\title{getAudienceById
Wrapper for Get Audience by ID API call (http://developers.adobetarget.com/api/#get-audience-by-id)}
\usage{
getAudienceById(audienceId)
}
\value{
data.frame of all activities
}
\description{
getAudienceById
Wrapper for Get Audience by ID API call (http://developers.adobetarget.com/api/#get-audience-by-id)
}
\examples{
getAudienceById("12345")
}
