% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listActivities.R
\name{listActivities}
\alias{listActivities}
\title{listActivities
Wrapper for List Activities API call (http://developers.adobetarget.com/api/#list-activities)}
\usage{
listActivities()
}
\value{
data.frame of all activities
}
\description{
listActivities
Wrapper for List Activities API call (http://developers.adobetarget.com/api/#list-activities)
}
\examples{
listActivities()
}
