% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAudiences.R
\name{listAudiences}
\alias{listAudiences}
\title{listAudiences
Wrapper for List Audiences API call (http://developers.adobetarget.com/api/#list-audiences)}
\usage{
listAudiences()
}
\value{
data.frame of all activities
}
\description{
listAudiences
Wrapper for List Audiences API call (http://developers.adobetarget.com/api/#list-audiences)
}
\examples{
listAudiences()
}
